Don't call unlink on a disposed property from PaintObserver, fixes https://github.com/phetsims/scenery/issues/692

+++ js/display/PaintObserver.js
@@ -265,7 +265,9 @@ define( function( require ) {
       if ( this.secondaryListenedCounts[ index ] === 0 ) {
         this.secondaryListenedProperties.splice( index, 1 );
         this.secondaryListenedCounts.splice( index, 1 );
-        property.unlink( this.updateSecondaryListener );
+        if ( !property.isDisposed ) {
+          property.unlink( this.updateSecondaryListener );
+        }
       }
     },
